[move]: move design_system to utilities
