+++ lib/src/components/greenfield_content_widget.dart
@@ -2,9 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:green_field/src/enums/feature_type.dart';
 import 'package:green_field/src/utilities/extensions/theme_data_extension.dart';
 import 'package:green_field/src/model/recruit.dart';
-
-import '../design_system/app_icons.dart';
-import '../design_system/app_texts.dart';
+import '../utilities/design_system/app_texts.dart';
+import '../utilities/design_system/app_icons.dart';
 
 class GreenFieldContentWidget extends StatelessWidget {
   final FeatureType? featureType;
