+++ lib/src/components/greenfield_list.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:green_field/src/design_system/app_icons.dart';
-import 'package:green_field/src/design_system/app_texts.dart';
+import '../utilities/design_system/app_texts.dart';
+import '../utilities/design_system/app_icons.dart';
 import 'package:green_field/src/utilities/extensions/theme_data_extension.dart';
 
 class GreenFieldList extends StatelessWidget {
