+++ lib/src/components/greenfield_text_field.dart
@@ -2,7 +2,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:green_field/src/enums/feature_type.dart';
 import 'package:green_field/src/utilities/extensions/theme_data_extension.dart';
-import '../design_system/app_texts.dart';
+import '../utilities/design_system/app_texts.dart';
 
 class GreenFieldTextField extends StatefulWidget {
   final FeatureType type;
