+++ lib/src/utilities/extensions/theme_data_extension.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import '../../design_system/app_colors.dart';
-import '../../design_system/app_texts.dart';
+import '../design_system/app_texts.dart';
+import '../design_system/app_colors.dart';
 
 
 extension ThemeDataExtended on ThemeData {
