+++ lib/src/views/campus/campus_operating_section.dart
@@ -3,9 +3,9 @@ import 'package:flutter/material.dart';
 import 'package:green_field/src/utilities/extensions/theme_data_extension.dart';
 import 'package:url_launcher/url_launcher.dart';
 
-import '../../design_system/app_icons.dart';
-import '../../design_system/app_texts.dart';
 import '../../model/campus.dart';
+import '../../utilities/design_system/app_icons.dart';
+import '../../utilities/design_system/app_texts.dart';
 
 class CampusOperatingSection extends StatelessWidget {
   const CampusOperatingSection({super.key});
