+++ lib/src/views/campus/campus_view.dart
@@ -5,7 +5,7 @@ import 'package:green_field/src/views/campus/camus_floor_section.dart';
 import 'package:green_field/src/views/campus/camus_map_section.dart';
 import 'package:green_field/src/views/campus/campus_operating_section.dart';
 
-import '../../design_system/app_texts.dart';
+import '../../utilities/design_system/app_texts.dart';
 
 class CampusView extends StatefulWidget {
   const CampusView({super.key});
