+++ lib/src/views/home/sections/notice_carousel_section.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:green_field/src/utilities/extensions/theme_data_extension.dart';
 
-import '../../../design_system/app_texts.dart';
+import '../../../utilities/design_system/app_texts.dart';
 import '../../../viewmodels/notice_view_model.dart';
 
 class NoticeCarouselSection extends StatefulWidget {
