+++ lib/src/views/home/sections/top_liked_posts_section.dart
@@ -4,8 +4,8 @@ import 'package:go_router/go_router.dart';
 import 'package:green_field/src/utilities/extensions/theme_data_extension.dart';
 import 'package:intl/intl.dart';
 
-import 'package:green_field/src/design_system/app_icons.dart';
-import '../../../design_system/app_texts.dart';
+import '../../../utilities/design_system/app_texts.dart';
+import '../../../utilities/design_system/app_icons.dart';
 import '../../../model/post.dart';
 import '../../../viewmodels/post_view_model.dart';
