+++ lib/src/views/login/login_view.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
-import 'package:green_field/src/design_system/app_colors.dart';
-import 'package:green_field/src/design_system/app_icons.dart';
-import 'package:green_field/src/design_system/app_texts.dart';
+import '../../utilities/design_system/app_colors.dart';
+import '../../utilities/design_system/app_icons.dart';
+import '../../utilities/design_system/app_texts.dart';
 import 'package:green_field/src/enums/login_type.dart';
 import 'package:green_field/src/utilities/extensions/theme_data_extension.dart';
