+++ lib/src/views/login/onboarding_view.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:green_field/src/components/greenfield_campus_picker.dart';
 import 'package:green_field/src/components/greenfield_confirm_button.dart';
-import 'package:green_field/src/design_system/app_colors.dart';
-import 'package:green_field/src/design_system/app_texts.dart';
+import '../../utilities/design_system/app_colors.dart';
+import '../../utilities/design_system/app_texts.dart';
 import 'package:green_field/src/utilities/extensions/theme_data_extension.dart';
 
 class OnboardingView extends StatelessWidget {
