+++ lib/src/views/recruitment/picker/recruit_picker_modal.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:green_field/src/design_system/app_texts.dart';
+import '../../../utilities/design_system/app_texts.dart';
 import 'package:green_field/src/enums/recruit_setting_type.dart';
 import 'package:green_field/src/utilities/extensions/theme_data_extension.dart';
