+++ lib/src/views/recruitment/recruit_edit_view.dart
@@ -5,8 +5,7 @@ import 'package:green_field/src/components/greenfield_app_bar.dart';
 import 'package:green_field/src/components/greenfield_edit_section.dart'; // Assuming you have a GreenFieldEditSection
 import 'package:green_field/src/enums/feature_type.dart';
 import 'package:green_field/src/utilities/extensions/theme_data_extension.dart';
-
-import '../../design_system/app_texts.dart'; // Assuming you have a Recruit model
+import '../../utilities/design_system/app_texts.dart';
 
 class RecruitEditView extends StatefulWidget {
   const RecruitEditView({super.key});
