+++ lib/src/views/recruitment/recruit_setting_section.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:green_field/src/utilities/extensions/theme_data_extension.dart';
 import 'package:green_field/src/views/recruitment/picker/recruit_picker_modal.dart';
 
-import '../../design_system/app_texts.dart';
+import '../../utilities/design_system/app_texts.dart';
 import '../../enums/recruit_setting_type.dart';
 
 class RecruitSettingSection extends StatefulWidget {
